pull grid click handling up into Crossword component
